Update TicTacToe.cpp
Improved the performance of program by adding constant reference to your function since passing by value can be expensive. Fixed some of the formatting on your code to make it more readable. Added Getlines to your main because that is the "best practice" way of obtaining strings from the user like names